chore: pricing v4 (#14336)
* chore: pricing v4

* fix: test

* chore: remove unused code

* chore: update

Former-commit-id: 2701f3d662f67a8649412d3935dbcfc7aaeb42b2

+++ backend/api/v1/environment_service.go
@@ -74,7 +74,7 @@ func (s *EnvironmentService) CreateEnvironment(ctx context.Context, request *v1p
 		return nil, status.Error(codes.Internal, err.Error())
 	}
 	maximumEnvironmentLimit := s.licenseService.GetPlanLimitValue(ctx, enterprise.PlanLimitMaximumEnvironment)
-	if int64(len(environments)) >= maximumEnvironmentLimit {
+	if len(environments) >= maximumEnvironmentLimit {
 		return nil, status.Errorf(codes.ResourceExhausted, "current plan can create up to %d environments.", maximumEnvironmentLimit)
 	}
